third_party/blink/web_tests/flag-specific/highdpi/virtual/container-queries/wpt_internal/css/css-conditional/container-queries/at-container-005-expected.txt

This is a testharness.js-based test.
FAIL Pseudo-elements ::before and ::after respond to container size changes assert_equals: expected "none" but got "\"before\""
FAIL Pseudo-element ::marker responds to container size changes assert_equals: expected "rgb(0, 0, 0)" but got "rgb(0, 128, 0)"
Harness: the test ran to completion.
